Update and create new interfaces and handle functions

+++ screens/admin/ProductScreen/ProductScreen.tsx
@@ -28,15 +28,9 @@ const ProductScreen = ({navigation}: any) => {
     const [dish, setDish] = useState<DishItem[]>([]);
     const [isLoading, setLoading] = useState(true);
 
-    const data = [
-        {"id": 1, "name": "Gà hấp", "price": 175000.0, "imageUrl": "http://192.168.0.104:8888/uploads/dishs/1716959172443_hngtrchanh.png", "categoryId": 1},
-        {"id": 2, "name": "Gà chiên", "price": 175000.0, "imageUrl": "http://192.168.0.104:8888/uploads/dishs/1716959172443_hngtrchanh.png", "categoryId": 2},
-        {"id": 3, "name": "Gà xào", "price": 175000.0, "imageUrl": "http://192.168.0.104:8888/uploads/dishs/1716959172443_hngtrchanh.png", "categoryId": 3},
-    ];
-
     const getAPI = async () => {
         try {
-            const response = await fetch(`${CONFIG.API_BASE_URL}/dishs`);
+            const response = await fetch(`${CONFIG.API_BASE_URL}/dishes`);
             const data = await response.json();
             setDish(data.data);
         } catch (error) {
@@ -61,24 +55,39 @@ const ProductScreen = ({navigation}: any) => {
         navigation.navigate('AddDish');
     }
 
-    const CategoryDeleteClick = async (id: number) => {
-        /*try {
-            const response = await fetch(`http://192.168.0.105:8888/api/v1/categorys/${id}`, {
-                method: 'DELETE',
-            });
-            const data = await response.json();
-            if (response.ok) {
-                Alert.alert("Thông báo", data.message);
-                // Cập nhật danh sách danh mục sau khi xóa thành công
-                const updatedCategories = category.filter(category => category.id !== id);
-                setCategory(updatedCategories);
-            } else {
-                Alert.alert("Thông báo", data.message);
-            }
-        } catch (error) {
-            console.log(error);
-            Alert.alert("Thông báo", "Có lỗi xảy ra!");
-        }*/
+    const ProductDeleteClick = async (id: number) => {
+        Alert.alert(
+            "Xác nhận",
+            "Bạn có chắc chắn muốn xóa danh mục này?",
+            [
+                {
+                    text: "Hủy",
+                    style: "cancel"
+                },
+                {
+                    text: "Xóa",
+                    onPress: async () => {
+                        try {
+                            const response = await fetch(`${CONFIG.API_BASE_URL}/dishes/${id}`, {
+                                method: 'DELETE',
+                            });
+                            const data = await response.json();
+                            if (response.ok) {
+                                Alert.alert("Thông báo", data.message);
+                                // Cập nhật danh sách danh mục sau khi xóa thành công
+                                const updatedDishes = dish.filter(dish => dish.id !== id);
+                                setDish(updatedDishes);
+                            } else {
+                                Alert.alert("Thông báo", data.message);
+                            }
+                        } catch (error) {
+                            console.log(error);
+                            Alert.alert("Thông báo", "Có lỗi xảy ra!");
+                        }
+                    }
+                }
+            ]
+        );
     };
 
     const DishUpdateClick = (item: any) => {
@@ -95,7 +104,7 @@ const ProductScreen = ({navigation}: any) => {
                 ) : (
                     <FlatList
                         scrollEnabled={false}
-                        data={data}
+                        data={dish}
                         renderItem={({item}: any) => (
                             <View style={styles.item}>
                                 <Image source={{uri: item.imageUrl}} style={styles.image}/>
@@ -104,7 +113,7 @@ const ProductScreen = ({navigation}: any) => {
                                     <TouchableOpacity onPress={() => DishUpdateClick(item)}>
                                         <Icon name={'update'} style={styles.icon}/>
                                     </TouchableOpacity>
-                                    <TouchableOpacity onPress={() => CategoryDeleteClick(item.id)}>
+                                    <TouchableOpacity onPress={() => ProductDeleteClick(item.id)}>
                                         <Icon name={'delete'} style={styles.icon}/>
                                     </TouchableOpacity>
                                 </View>
